Route stats page (#65)
Implement a route stats page to show additional info per route. Route info is determined by URL parameter. The page is accessed from the navbar or by the "More Stats!" button in the map modal. There is a search bar on the page to change which route you are looking at.